// src/redux/types/customer/actionName.ts
export enum CustomerActionName {
  START_CUSTOMERS_ACTION = 'START_CUSTOMERS_ACTION',
  FETCH_CUSTOMERS = 'FETCH_CUSTOMERS',
  ERROR_CUSTOMERS_ACTION = 'ERROR_CUSTOMERS_ACTION',
  FETCH_CUSTOMERS_BY_GENDER = 'FETCH_CUSTOMERS_BY_GENDER',
  FETCH_CUSTOMERS_BY_AGE = 'FETCH_CUSTOMERS_BY_AGE',
  FETCH_CUSTOMERS_BY_NAT = 'FETCH_CUSTOMERS_BY_NAT'
}